requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
aiofiles>=23.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
